% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addPositionID}
\alias{addPositionID}
\title{Add position id, ie. chrX+231231 to GRange object.}
\usage{
addPositionID(gr)
}
\arguments{
\item{GRange}{Grange object.}
}
\description{
Add position id, ie. chrX+231231 to GRange object.
}
