% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{annotateIntSites}
\alias{annotateIntSites}
\title{Calculate nearest genomic and onocogene features.}
\usage{
annotateIntSites(sites, CPUs = 20, genomeFileMap = NULL,
  oncoGeneListSide = "either", lymphomaGenesListSide = "either")
}
\value{
GRange object updated with genomic feature annotations.
}
\description{
Calculate nearest genomic and onocogene features.
}
