% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredTools.R
\name{calcReplicateAbundances}
\alias{calcReplicateAbundances}
\title{Merge replicate samples and calculate abundances}
\usage{
calcReplicateAbundances(gr, method = "fragments")
}
\arguments{
\item{gr}{A GRange object of intSites.}

\item{method}{Method ('fragments' or 'reads') to use to determine clonal abudnace.}
}
\value{
A GRange object of intSites where sequence ranges hanve been reduced to their start
positions, relicate samples have been combined into single samples and a clonal abundance
column has been added.
}
\description{
Merge each replicated sample {sampleName} into single samples {GTSP} and calculate clonal
abundances {estAbund}.
}
