% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calcSampleAbunds}
\alias{calcSampleAbunds}
\title{Estimate abundances by number of unique fragments associated with integration positions for each sample.}
\usage{
calcSampleAbunds(f)
}
\arguments{
\item{f}{GRange object of standardized genomic fragments.}
}
\value{
GRange object of integration positions including estimated abundances.
}
\description{
Estimate abundances by number of unique fragments associated with integration positions for each sample.
}
