% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collapseReplicatesCalcAbunds}
\alias{collapseReplicatesCalcAbunds}
\title{Collapse technical replicates and estimate abundances by number of unique fragments associated with integration positions.}
\usage{
collapseReplicatesCalcAbunds(f)
}
\arguments{
\item{f}{GRange object of standardized genomic fragments.}
}
\value{
GRange object of integration positions including estimated abundances.
}
\description{
Collapse technical replicates and estimate abundances by number of unique fragments associated with integration positions.
}
