% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicAnnotation.R
\name{createIntUCSCTrack}
\alias{createIntUCSCTrack}
\title{Annotate genomic ranges}
\usage{
createIntUCSCTrack(d, abundCuts = c(5, 10, 50),
  posColors = c("#8C9DFF", "#6768E3", "#4234C7", "#1D00AB"),
  negColors = c("#FF8C8C", "#E35D5D", "#C72E2E", "#AB0000"),
  title = "intSites", outputFile = "track.ucsc", visibility = 1,
  position = "chr7:127471196-127495720", padSite = 0, siteLabel = NA)
}
\arguments{
\item{d}{Data fram containing genomic ranges to be added to UCSC track.}

\item{abundCuts}{Cut points for estimated abundance (estAbund) values.}

\item{posColors}{Color codes for binned abundances (positive integrations).}

\item{negColors}{Color codes for binned abundances (positive integrations).}

\item{title}{Track title.}

\item{outputFile}{Track output file.}

\item{visibility}{Track default visibility (0 - hide, 1 - dense, 2 - full, 3 - pack, and 4 - squish).}

\item{position}{Deafult track position.}

\item{padSite}{Number of NTs to pad sites with for increased visibility.}

\item{siteLabel}{Text to appear next to sites, ie. 'Patient X, chr12+1052325'.}
}
\value{
Nothing.
}
\description{
Create a dataframe of sequencing run ids associated with 1 or more patient identifiers.
}
