% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{defaultGenomeFileMappings}
\alias{defaultGenomeFileMappings}
\title{Create a list of default file mappings.}
\usage{
defaultGenomeFileMappings()
}
\value{
List of default gt23 file mappings.
}
\description{
An empty character vector object named 'none' is available for instances where 
gene lists are not available.
}
