% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredDB.R
\name{expandTimePoints}
\alias{expandTimePoints}
\title{Expand character time points to days and months.}
\usage{
expandTimePoints(tps)
}
\arguments{
\item{tps}{Character vector of timepoints, ie. c('D0', 'd10', 'M5', '-d20').}
}
\value{
Data frame with days and months.
}
\description{
Expand character time points to days and months.
}
