% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredDB.R
\name{getDBgenomicFragments}
\alias{getDBgenomicFragments}
\title{Retrieve intSite data}
\usage{
getDBgenomicFragments(samples, sampleDB.group, intSiteDB.group)
}
\arguments{
\item{samples}{Vector of samples IDs for which to retrieve intSites.}

\item{sampleDB.group}{Sample database connection group.}

\item{intSiteDB.group}{IntSite database connection group.}
}
\value{
A GRange object of intSites.
}
\description{
Retrieve inSite data for either a number of patients, sequencing runs or samples.
}
