% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredDB.R
\name{getIntSiteData}
\alias{getIntSiteData}
\title{Retrieve intSite data}
\usage{
getIntSiteData(sampleDB.group, intSiteDB.group, patients = NULL,
  samples = NULL, roundTimePoints = TRUE, runIDs = NULL)
}
\arguments{
\item{sampleDB.group}{Sample database connection group}

\item{intSiteDB.group}{IntSite database connection group}

\item{patients}{Vector of patient IDs for which to retrieve intSites.}

\item{samples}{Vector of samples IDs for which to retrieve intSites.}

\item{roundTimePoints}{Logical determining if time points should be rounded.}

\item{runIDs}{Vector of sequencing run ids for which to retrieve intSites.}
}
\value{
A GRange object of intSites.
}
\description{
Retrieve inSite data for either a number of patients, sequencing runs or samples.
}
