% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredDB.R
\name{getPatientRunDetails}
\alias{getPatientRunDetails}
\title{Retrieve patient sequencing run data}
\usage{
getPatientRunDetails(patients,
  specimenManagementDBgroup = "specimen_management",
  intsitesDBgroup = "intsites_miseq")
}
\arguments{
\item{patients}{A vector of patient identifiers.}

\item{specimenManagementDBgroup}{MySQL database group for specimen management database.}

\item{intsitesDBgroup}{MySQL database group for intSite database.}
}
\value{
A data frame with both run ids and run dates.
}
\description{
Create a dataframe of sequencing run ids associated with 1 or more patient identifiers.
}
