% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredDB.R
\name{getRunDetails}
\alias{getRunDetails}
\title{Retrieve sequencing run patient details}
\usage{
getRunDetails(runID, specimenManagementDBgroup = "specimen_management",
  intsitesDBgroup = "intsites_miseq")
}
\arguments{
\item{runID}{A single MiSeq sequencing run id.}

\item{specimenManagementDBgroup}{MySQL database group for specimen management database.}

\item{intsitesDBgroup}{MySQL database group for intSite database.}
}
\value{
A data frame detailing patients and sample details.
}
\description{
Create a dataframe of detailing a specific sequencing run.
}
