% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredDB.R
\name{getSampleSheetDetails}
\alias{getSampleSheetDetails}
\title{Summarize samples in sequencer sample sheet file.}
\usage{
getSampleSheetDetails(sampleSheetFile,
  specimenManagementDBgroup = "specimen_management")
}
\value{
Data frame including Trial, Patient, CellType, Timepoint fields.
}
\description{
Sample sheets are part of the Bushman group INSPIIRED pipeline.
}
