% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{intSiteDistributionPlot}
\alias{intSiteDistributionPlot}
\title{Create a integration site location plot.}
\usage{
intSiteDistributionPlot(gr, chromosomeLengths, alpha = 0.8,
  siteColor = "black")
}
\arguments{
\item{gr}{GRange object.}

\item{chromosomeLengths}{List with chromosome names which stores chromosomal lengths.}

\item{alpha}{Alpha value of integration tick marks.}

\item{siteColor}{Color name or hex code for integration tick marks.}
}
\description{
Create a integration site location plot.
}
