% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{numShortHand}
\alias{numShortHand}
\title{Abbreviate numbers with scientific suffixes.}
\usage{
numShortHand(number, rounding = F, digits = ifelse(rounding, NA, 2))
}
\arguments{
\item{number}{Floating point or integer.}

\item{rounding}{Boolean, round provided number if TRUE.}

\item{digits}{Number of significant digits to round number to if rounding set to TRUE.}
}
\description{
Abbreviate numbers with scientific suffixes.
}
