% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{scanStats}
\alias{scanStats}
\title{Run geneRxCluster scan statistics.
Bioinformatics. 2014 Jun 1;30(11):1493-500.}
\usage{
scanStats(gr1, gr2, gr1.label = "A", gr2.label = "B",
  kvals = "15L:30L", nperm = 100,
  cutpt.tail.expr = "critVal.target(k, n, target = 5, posdiff = x)",
  cutpt.filter.expr = "as.double(apply(x, 2, median, na.rm = TRUE))")
}
\arguments{
\item{kvals:}{The interpretation of kvals = 2L:5L is that genomic windows will be drawn for every consecutive groups of integration sites with group sizes of 2, 3, 4, and 5 integration sites.}
}
\value{
Data frame of identified clusters with counts and statistics where target.min is the smallest nominal False Discoveries Expected for each cluster.
}
\description{
Run geneRxCluster scan statistics.
Bioinformatics. 2014 Jun 1;30(11):1493-500.
}
