% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspiiredTools.R
\name{standardizeSeqRanges.malani}
\alias{standardizeSeqRanges.malani}
\title{Standardize sequence ranges with Malani method}
\usage{
standardizeSeqRanges.malani(gr, standardizeStart = TRUE,
  standardizeEnd = TRUE)
}
\arguments{
\item{gr}{A GRange object of intSites.}

\item{standardizeStart}{Boolean which determines if sequence range start positions should be standardized.}

\item{standardizeEnd}{Boolean which determines if sequence range end positions should be standardized.}
}
\value{
A GRange object of intSites where sequence ranges have been standardized.
}
\description{
Standardize sequence ranges with Malani method
}
