% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stdIntSiteFragments}
\alias{stdIntSiteFragments}
\title{Standardize genomic fragments from intSite experiments using the gintools package.}
\usage{
stdIntSiteFragments(frags, CPUs = 10, countsCol = "reads")
}
\arguments{
\item{frags}{GenomicRange object containing genomic fragments where fragments with positive strands 
are assumed to have their integration positions on the left and break points on the right (opposite assumed from negative strand).}

\item{CPUs}{Number of CPUs to use for break point standardizations.}

\item{countsCol}{Name of meta data column containing the number of times fragments were observed.}
}
\description{
Standardize genomic fragments from intSite experiments using the gintools package.
}
