% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{tuneScanStatistics}
\alias{tuneScanStatistics}
\title{Run a number of combination of settings for geneRxCluster scan statistics.
Bioinformatics. 2014 Jun 1;30(11):1493-500.}
\usage{
tuneScanStatistics(A, B, CPUs = 25, minWindow = 10, maxWindow = 30,
  windowStep = 5, minProb = 0.75, maxProb = 0.95, probStep = 0.025,
  minTarget = 10, maxTarget = 25, targetStep = 5)
}
\value{
Data frame of clusters statistics from each combination of paramteres tested.
}
\description{
Run a number of combination of settings for geneRxCluster scan statistics.
Bioinformatics. 2014 Jun 1;30(11):1493-500.
}
