% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{validateGenomeFileMap}
\alias{validateGenomeFileMap}
\title{Validate that each named data object in genome file mapping list is present in the gt23 package.}
\usage{
validateGenomeFileMap(x)
}
\value{
Error if one or more named data objects are not present.
}
\description{
The special object 'none' is allowed which is an empty character vector object.
}
