% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicAnnotation.R
\name{nearestGenomicFeature}
\alias{nearestGenomicFeature}
\title{Annotate genomic ranges}
\usage{
nearestGenomicFeature(query, subject = NULL, subject.exons = NULL,
  side = "either", geneList = NULL)
}
\arguments{
\item{query}{GRange object containing ranges from a genome supported by the gt23 package.}

\item{subject}{gt23 genome specific gene boundary object.}

\item{subject.exons}{gt23 genome specific gene exon object.}

\item{side}{Which query boundary should be tested ('5p', '3p', 'either').}

\item{geneList}{Case insensitive list of gene names to filter nearest features against.}
}
\value{
An updated query GRange object.
}
\description{
Create a dataframe of sequencing run ids associated with 1 or more patient identifiers.
}
